Move diff_on_edges_torus_numpy to distance_array_ns
