+++ model/common/src/icon4py/model/common/math/projection.py
@@ -48,28 +48,3 @@ def gnomonic_proj(
         - array_ns.sin(lat_c) * array_ns.cos(lat) * array_ns.cos(lon - lon_c)
     )
     return array_ns.column_stack((x, y)) * sphere_radius
-
-
-def diff_on_edges_torus_numpy(
-    *,
-    cc_cv_x: float,
-    cc_cv_y: float,
-    cc_cell_x: float,
-    cc_cell_y: float,
-    domain_length: float,
-    domain_height: float,
-) -> tuple[float, float]:
-    if abs(cc_cell_x - cc_cv_x) <= 0.5 * domain_length:
-        x1 = cc_cell_x
-    elif cc_cv_x > cc_cell_x:
-        x1 = cc_cell_x + domain_length
-    else:
-        x1 = cc_cell_x - domain_length
-
-    if abs(cc_cell_y - cc_cv_y) <= 0.5 * domain_height:
-        y1 = cc_cell_y
-    elif cc_cv_y > cc_cell_y:
-        y1 = cc_cell_y + domain_height
-    else:
-        y1 = cc_cell_y - domain_height
-    return x1, y1
